Reduce codegen.ts file size by extracting modules [L] (#399)
