+++ src/compiler/codegen.ts
@@ -1,10 +1,7 @@
 import { DEFAULT_CONFIG } from "../config/defaults.ts";
 import {
-  ADDRESS_LITERAL_RE,
   type Expression,
-  type IfStatement,
   type NatSpec,
-  type RevertStatement,
   type SkittlesConstructor,
   type SkittlesContract,
   type SkittlesContractInterface,
@@ -15,809 +12,50 @@ import {
   SkittlesTypeKind,
   type SkittlesVariable,
   type SolidityConfig,
-  type SourceMapping,
   type Statement,
   type Visibility,
 } from "../types/index.ts";
 import { cctx } from "./codegen-context.ts";
-import {
-  CHAR_0,
-  CHAR_A,
-  CHAR_a,
-  CHAR_SPACE,
-  CHAR_Z,
-  CHAR_z,
-  ERR_START_OUT_OF_BOUNDS,
-} from "./constants.ts";
 import { walkStatements } from "./walker.ts";
+
+// Sub-module imports
+import {
+  collectLocalVarNames,
+  pickNewName,
+  renameInExpression,
+  renameInStatements,
+  scopeAwareRenameBlock,
+  resolveShadowedLocals,
+} from "./codegen/rename.ts";
+import { emitHelperFunctions, hasAncestorOrigin } from "./codegen/helpers.ts";
+import {
+  generateType,
+  generateParamType,
+  generateCalldataParamType,
+  generateExpression,
+} from "./codegen/expressions.ts";
+import {
+  generateStatement,
+  isRequirePattern,
+  negateExpression,
+} from "./codegen/statements.ts";
+import {
+  buildSourceMap,
+  computeAncestors,
+  getFunctionKey,
+} from "./codegen/sourcemap.ts";
+
+// Re-exports for public API
 export type { CodegenContext } from "./codegen-context.ts";
 export { resetCodegenContext } from "./codegen-context.ts";
-
-// Registry of helper function names to their Solidity code lines.
-// Each entry maps a helper name to a thunk returning its code lines
-// (thunks are used so that constant references are evaluated at call time).
-const HELPER_REGISTRY: Array<{ name: string; lines: () => string[] }> = [
-  {
-    name: "_min",
-    lines: () => [
-      "    function _min(uint256 a, uint256 b) internal pure returns (uint256) {",
-      "        return a < b ? a : b;",
-      "    }",
-    ],
-  },
-  {
-    name: "_max",
-    lines: () => [
-      "    function _max(uint256 a, uint256 b) internal pure returns (uint256) {",
-      "        return a > b ? a : b;",
-      "    }",
-    ],
-  },
-  {
-    name: "_sqrt",
-    lines: () => [
-      "    function _sqrt(uint256 x) internal pure returns (uint256) {",
-      "        if (x == 0) return 0;",
-      "        uint256 z = (x + 1) / 2;",
-      "        uint256 y = x;",
-      "        while (z < y) {",
-      "            y = z;",
-      "            z = (x / z + z) / 2;",
-      "        }",
-      "        return y;",
-      "    }",
-    ],
-  },
-  {
-    name: "_charAt",
-    lines: () => [
-      "    function _charAt(string memory str, uint256 index) internal pure returns (string memory) {",
-      "        bytes memory strBytes = bytes(str);",
-      "        require(index < strBytes.length);",
-      "        bytes memory result = new bytes(1);",
-      "        result[0] = strBytes[index];",
-      "        return string(result);",
-      "    }",
-    ],
-  },
-  {
-    name: "_substring",
-    lines: () => [
-      "    function _substring(string memory str, uint256 start, uint256 end) internal pure returns (string memory) {",
-      "        bytes memory strBytes = bytes(str);",
-      "        require(start <= end && end <= strBytes.length);",
-      "        bytes memory result = new bytes(end - start);",
-      "        for (uint256 i = start; i < end; i++) {",
-      "            result[i - start] = strBytes[i];",
-      "        }",
-      "        return string(result);",
-      "    }",
-    ],
-  },
-  {
-    name: "_toLowerCase",
-    lines: () => [
-      "    function _toLowerCase(string memory str) internal pure returns (string memory) {",
-      "        bytes memory strBytes = bytes(str);",
-      "        bytes memory result = new bytes(strBytes.length);",
-      "        for (uint256 i = 0; i < strBytes.length; i++) {",
-      "            uint8 c = uint8(strBytes[i]);",
-      `            if (c >= ${CHAR_A} && c <= ${CHAR_Z}) {`,
-      `                result[i] = bytes1(c + ${CHAR_SPACE});`,
-      "            } else {",
-      "                result[i] = strBytes[i];",
-      "            }",
-      "        }",
-      "        return string(result);",
-      "    }",
-    ],
-  },
-  {
-    name: "_toUpperCase",
-    lines: () => [
-      "    function _toUpperCase(string memory str) internal pure returns (string memory) {",
-      "        bytes memory strBytes = bytes(str);",
-      "        bytes memory result = new bytes(strBytes.length);",
-      "        for (uint256 i = 0; i < strBytes.length; i++) {",
-      "            uint8 c = uint8(strBytes[i]);",
-      `            if (c >= ${CHAR_a} && c <= ${CHAR_z}) {`,
-      `                result[i] = bytes1(c - ${CHAR_SPACE});`,
-      "            } else {",
-      "                result[i] = strBytes[i];",
-      "            }",
-      "        }",
-      "        return string(result);",
-      "    }",
-    ],
-  },
-  {
-    name: "_startsWith",
-    lines: () => [
-      "    function _startsWith(string memory str, string memory prefix) internal pure returns (bool) {",
-      "        bytes memory strBytes = bytes(str);",
-      "        bytes memory prefixBytes = bytes(prefix);",
-      "        if (prefixBytes.length > strBytes.length) return false;",
-      "        for (uint256 i = 0; i < prefixBytes.length; i++) {",
-      "            if (strBytes[i] != prefixBytes[i]) return false;",
-      "        }",
-      "        return true;",
-      "    }",
-    ],
-  },
-  {
-    name: "_endsWith",
-    lines: () => [
-      "    function _endsWith(string memory str, string memory suffix) internal pure returns (bool) {",
-      "        bytes memory strBytes = bytes(str);",
-      "        bytes memory suffixBytes = bytes(suffix);",
-      "        if (suffixBytes.length > strBytes.length) return false;",
-      "        uint256 offset = strBytes.length - suffixBytes.length;",
-      "        for (uint256 i = 0; i < suffixBytes.length; i++) {",
-      "            if (strBytes[offset + i] != suffixBytes[i]) return false;",
-      "        }",
-      "        return true;",
-      "    }",
-    ],
-  },
-  {
-    name: "_trim",
-    lines: () => [
-      "    function _trim(string memory str) internal pure returns (string memory) {",
-      "        bytes memory strBytes = bytes(str);",
-      "        uint256 start = 0;",
-      "        uint256 end = strBytes.length;",
-      `        while (start < end && uint8(strBytes[start]) == ${CHAR_SPACE}) { start++; }`,
-      `        while (end > start && uint8(strBytes[end - 1]) == ${CHAR_SPACE}) { end--; }`,
-      "        bytes memory result = new bytes(end - start);",
-      "        for (uint256 i = start; i < end; i++) {",
-      "            result[i - start] = strBytes[i];",
-      "        }",
-      "        return string(result);",
-      "    }",
-    ],
-  },
-  {
-    name: "_split",
-    lines: () => [
-      "    function _split(string memory str, string memory delimiter) internal pure returns (string[] memory) {",
-      "        bytes memory strBytes = bytes(str);",
-      "        bytes memory delimBytes = bytes(delimiter);",
-      "        require(delimBytes.length > 0);",
-      "        uint256 __sk_count = 1;",
-      "        for (uint256 i = 0; i + delimBytes.length <= strBytes.length; i++) {",
-      "            bool found = true;",
-      "            for (uint256 j = 0; j < delimBytes.length; j++) {",
-      "                if (strBytes[i + j] != delimBytes[j]) { found = false; break; }",
-      "            }",
-      "            if (found) { __sk_count++; i += delimBytes.length - 1; }",
-      "        }",
-      "        string[] memory parts = new string[](__sk_count);",
-      "        uint256 partIndex = 0;",
-      "        uint256 start = 0;",
-      "        for (uint256 i = 0; i + delimBytes.length <= strBytes.length; i++) {",
-      "            bool found = true;",
-      "            for (uint256 j = 0; j < delimBytes.length; j++) {",
-      "                if (strBytes[i + j] != delimBytes[j]) { found = false; break; }",
-      "            }",
-      "            if (found) {",
-      "                bytes memory part = new bytes(i - start);",
-      "                for (uint256 k = start; k < i; k++) { part[k - start] = strBytes[k]; }",
-      "                parts[partIndex++] = string(part);",
-      "                start = i + delimBytes.length;",
-      "                i += delimBytes.length - 1;",
-      "            }",
-      "        }",
-      "        bytes memory lastPart = new bytes(strBytes.length - start);",
-      "        for (uint256 k = start; k < strBytes.length; k++) { lastPart[k - start] = strBytes[k]; }",
-      "        parts[partIndex] = string(lastPart);",
-      "        return parts;",
-      "    }",
-    ],
-  },
-  {
-    name: "__sk_toString",
-    lines: () => [
-      "    function __sk_toString(uint256 value) internal pure returns (string memory) {",
-      '        if (value == 0) return "0";',
-      "        uint256 temp = value;",
-      "        uint256 digits;",
-      "        while (temp != 0) { digits++; temp /= 10; }",
-      "        bytes memory buffer = new bytes(digits);",
-      "        while (value != 0) {",
-      "            digits--;",
-      `            buffer[digits] = bytes1(uint8(${CHAR_0} + (value % 10)));`,
-      "            value /= 10;",
-      "        }",
-      "        return string(buffer);",
-      "    }",
-    ],
-  },
-  {
-    name: "_replace",
-    lines: () => [
-      "    function _replace(string memory str, string memory search, string memory replacement) internal pure returns (string memory) {",
-      "        bytes memory strBytes = bytes(str);",
-      "        bytes memory searchBytes = bytes(search);",
-      "        bytes memory replBytes = bytes(replacement);",
-      "        require(searchBytes.length > 0);",
-      "        for (uint256 i = 0; i + searchBytes.length <= strBytes.length; i++) {",
-      "            bool found = true;",
-      "            for (uint256 j = 0; j < searchBytes.length; j++) {",
-      "                if (strBytes[i + j] != searchBytes[j]) { found = false; break; }",
-      "            }",
-      "            if (found) {",
-      "                bytes memory result = new bytes(strBytes.length - searchBytes.length + replBytes.length);",
-      "                for (uint256 k = 0; k < i; k++) { result[k] = strBytes[k]; }",
-      "                for (uint256 k = 0; k < replBytes.length; k++) { result[i + k] = replBytes[k]; }",
-      "                for (uint256 k = i + searchBytes.length; k < strBytes.length; k++) { result[k - searchBytes.length + replBytes.length] = strBytes[k]; }",
-      "                return string(result);",
-      "            }",
-      "        }",
-      "        return str;",
-      "    }",
-    ],
-  },
-  {
-    name: "_replaceAll",
-    lines: () => [
-      "    function _replaceAll(string memory str, string memory search, string memory replacement) internal pure returns (string memory) {",
-      "        bytes memory strBytes = bytes(str);",
-      "        bytes memory searchBytes = bytes(search);",
-      "        bytes memory replBytes = bytes(replacement);",
-      "        require(searchBytes.length > 0);",
-      "        uint256 __sk_count = 0;",
-      "        for (uint256 i = 0; i + searchBytes.length <= strBytes.length; i++) {",
-      "            bool found = true;",
-      "            for (uint256 j = 0; j < searchBytes.length; j++) {",
-      "                if (strBytes[i + j] != searchBytes[j]) { found = false; break; }",
-      "            }",
-      "            if (found) { __sk_count++; i += searchBytes.length - 1; }",
-      "        }",
-      "        if (__sk_count == 0) return str;",
-      "        bytes memory result = new bytes(strBytes.length - (__sk_count * searchBytes.length) + (__sk_count * replBytes.length));",
-      "        uint256 idx = 0;",
-      "        for (uint256 i = 0; i < strBytes.length; ) {",
-      "            bool found = false;",
-      "            if (i + searchBytes.length <= strBytes.length) {",
-      "                found = true;",
-      "                for (uint256 j = 0; j < searchBytes.length; j++) {",
-      "                    if (strBytes[i + j] != searchBytes[j]) { found = false; break; }",
-      "                }",
-      "            }",
-      "            if (found) {",
-      "                for (uint256 k = 0; k < replBytes.length; k++) { result[idx++] = replBytes[k]; }",
-      "                i += searchBytes.length;",
-      "            } else {",
-      "                result[idx++] = strBytes[i];",
-      "                i++;",
-      "            }",
-      "        }",
-      "        return string(result);",
-      "    }",
-    ],
-  },
-];
-
-const SOLIDITY_VALUE_TYPES = new Set([
-  "uint256",
-  "int256",
-  "address",
-  "bool",
-  "bytes32",
-  "uint8",
-  "uint16",
-  "uint32",
-  "uint64",
-  "uint128",
-  "int8",
-  "int16",
-  "int32",
-  "int64",
-  "int128",
-  "bytes1",
-  "bytes2",
-  "bytes3",
-  "bytes4",
-]);
-
-// ============================================================
-// Shadowed local variable renaming
-// ============================================================
-
-function collectLocalVarNames(stmts: Statement[]): Set<string> {
-  const names = new Set<string>();
-  for (const s of stmts) {
-    switch (s.kind) {
-      case "variable-declaration":
-        names.add(s.name);
-        break;
-      case "tuple-destructuring":
-        for (const n of s.names) if (n !== null) names.add(n);
-        break;
-      case "if":
-        for (const n of collectLocalVarNames(s.thenBody)) names.add(n);
-        if (s.elseBody)
-          for (const n of collectLocalVarNames(s.elseBody)) names.add(n);
-        break;
-      case "for":
-        if (s.initializer?.kind === "variable-declaration")
-          names.add(s.initializer.name);
-        for (const n of collectLocalVarNames(s.body)) names.add(n);
-        break;
-      case "while":
-      case "do-while":
-        for (const n of collectLocalVarNames(s.body)) names.add(n);
-        break;
-      case "switch":
-        for (const c of s.cases)
-          for (const n of collectLocalVarNames(c.body)) names.add(n);
-        break;
-      case "try-catch":
-        if (s.returnVarName) names.add(s.returnVarName);
-        for (const n of collectLocalVarNames(s.successBody)) names.add(n);
-        for (const n of collectLocalVarNames(s.catchBody)) names.add(n);
-        break;
-    }
-  }
-  return names;
-}
-
-function pickNewName(original: string, taken: Set<string>): string {
-  let candidate = `_${original}`;
-  while (taken.has(candidate)) {
-    candidate = `_${candidate}`;
-  }
-  return candidate;
-}
-
-function renameInExpression(
-  expr: Expression,
-  renames: Map<string, string>
-): Expression {
-  switch (expr.kind) {
-    case "identifier": {
-      const newName = renames.get(expr.name);
-      return newName ? { ...expr, name: newName } : expr;
-    }
-    case "binary":
-      return {
-        ...expr,
-        left: renameInExpression(expr.left, renames),
-        right: renameInExpression(expr.right, renames),
-      };
-    case "unary":
-      return { ...expr, operand: renameInExpression(expr.operand, renames) };
-    case "assignment":
-      return {
-        ...expr,
-        target: renameInExpression(expr.target, renames),
-        value: renameInExpression(expr.value, renames),
-      };
-    case "call":
-      return {
-        ...expr,
-        callee: renameInExpression(expr.callee, renames),
-        args: expr.args.map((a) => renameInExpression(a, renames)),
-      };
-    case "conditional":
-      return {
-        ...expr,
-        condition: renameInExpression(expr.condition, renames),
-        whenTrue: renameInExpression(expr.whenTrue, renames),
-        whenFalse: renameInExpression(expr.whenFalse, renames),
-      };
-    case "property-access":
-      return { ...expr, object: renameInExpression(expr.object, renames) };
-    case "element-access":
-      return {
-        ...expr,
-        object: renameInExpression(expr.object, renames),
-        index: renameInExpression(expr.index, renames),
-      };
-    case "new":
-      return {
-        ...expr,
-        args: expr.args.map((a) => renameInExpression(a, renames)),
-      };
-    case "object-literal":
-      return {
-        ...expr,
-        properties: expr.properties.map((p) => ({
-          ...p,
-          value: renameInExpression(p.value, renames),
-        })),
-      };
-    case "tuple-literal":
-      return {
-        ...expr,
-        elements: expr.elements.map((e) => renameInExpression(e, renames)),
-      };
-    default:
-      return expr;
-  }
-}
-
-function renameInStatement(
-  stmt: Statement,
-  renames: Map<string, string>
-): Statement {
-  switch (stmt.kind) {
-    case "variable-declaration": {
-      const newName = renames.get(stmt.name) ?? stmt.name;
-      return {
-        ...stmt,
-        name: newName,
-        initializer: stmt.initializer
-          ? renameInExpression(stmt.initializer, renames)
-          : undefined,
-      };
-    }
-    case "tuple-destructuring": {
-      const newNames = stmt.names.map((n) =>
-        n === null ? null : (renames.get(n) ?? n)
-      );
-      return {
-        ...stmt,
-        names: newNames,
-        initializer: renameInExpression(stmt.initializer, renames),
-      };
-    }
-    case "return":
-      return {
-        ...stmt,
-        value: stmt.value ? renameInExpression(stmt.value, renames) : undefined,
-      };
-    case "expression":
-      return {
-        ...stmt,
-        expression: renameInExpression(stmt.expression, renames),
-      };
-    case "if":
-      return {
-        ...stmt,
-        condition: renameInExpression(stmt.condition, renames),
-        thenBody: renameInStatements(stmt.thenBody, renames),
-        elseBody: stmt.elseBody
-          ? renameInStatements(stmt.elseBody, renames)
-          : undefined,
-      };
-    case "for": {
-      let init = stmt.initializer;
-      if (init) {
-        init = renameInStatement(init, renames) as typeof init;
-      }
-      return {
-        ...stmt,
-        initializer: init,
-        condition: stmt.condition
-          ? renameInExpression(stmt.condition, renames)
-          : undefined,
-        incrementor: stmt.incrementor
-          ? renameInExpression(stmt.incrementor, renames)
-          : undefined,
-        body: renameInStatements(stmt.body, renames),
-      };
-    }
-    case "while":
-    case "do-while":
-      return {
-        ...stmt,
-        condition: renameInExpression(stmt.condition, renames),
-        body: renameInStatements(stmt.body, renames),
-      };
-    case "emit":
-      return {
-        ...stmt,
-        args: stmt.args.map((a) => renameInExpression(a, renames)),
-      };
-    case "revert":
-      return {
-        ...stmt,
-        message: stmt.message
-          ? renameInExpression(stmt.message, renames)
-          : undefined,
-        customErrorArgs: stmt.customErrorArgs?.map((a) =>
-          renameInExpression(a, renames)
-        ),
-      };
-    case "delete":
-      return { ...stmt, target: renameInExpression(stmt.target, renames) };
-    case "switch":
-      return {
-        ...stmt,
-        discriminant: renameInExpression(stmt.discriminant, renames),
-        cases: stmt.cases.map((c) => ({
-          ...c,
-          value: c.value ? renameInExpression(c.value, renames) : undefined,
-          body: renameInStatements(c.body, renames),
-        })),
-      };
-    case "try-catch": {
-      const newReturnVarName = stmt.returnVarName
-        ? (renames.get(stmt.returnVarName) ?? stmt.returnVarName)
-        : undefined;
-      return {
-        ...stmt,
-        call: renameInExpression(stmt.call, renames),
-        returnVarName: newReturnVarName,
-        successBody: renameInStatements(stmt.successBody, renames),
-        catchBody: renameInStatements(stmt.catchBody, renames),
-      };
-    }
-    case "console-log":
-      return {
-        ...stmt,
-        args: stmt.args.map((a) => renameInExpression(a, renames)),
-      };
-    default:
-      return stmt;
-  }
-}
-
-function renameInStatements(
-  stmts: Statement[],
-  renames: Map<string, string>
-): Statement[] {
-  return stmts.map((s) => renameInStatement(s, renames));
-}
-
-// Scope-aware renaming: renames are activated only when their declaration is
-// encountered, and inner-block declarations do not leak to outer scopes.
-
-function scopeAwareRenameBlock(
-  stmts: Statement[],
-  parentActive: Map<string, string>,
-  allRenames: Map<string, string>
-): Statement[] {
-  let active = new Map(parentActive);
-  const result: Statement[] = [];
-  for (const stmt of stmts) {
-    const out = scopeAwareRenameStmt(stmt, active, allRenames);
-    result.push(out.stmt);
-    active = out.active;
-  }
-  return result;
-}
-
-function scopeAwareRenameStmt(
-  stmt: Statement,
-  active: Map<string, string>,
-  allRenames: Map<string, string>
-): { stmt: Statement; active: Map<string, string> } {
-  switch (stmt.kind) {
-    case "variable-declaration": {
-      const rename = allRenames.get(stmt.name);
-      if (rename) {
-        // Rename the initializer with the parent scope (before declaration is active)
-        const init = stmt.initializer
-          ? renameInExpression(stmt.initializer, active)
-          : undefined;
-        const newActive = new Map(active);
-        newActive.set(stmt.name, rename);
-        return {
-          stmt: { ...stmt, name: rename, initializer: init },
-          active: newActive,
-        };
-      }
-      const init = stmt.initializer
-        ? renameInExpression(stmt.initializer, active)
-        : undefined;
-      // Suppress pre-activated rename when a body local re-declares the same name
-      // (used for ctor default params where renames start active via parentActive)
-      if (active.has(stmt.name) && !allRenames.has(stmt.name)) {
-        const newActive = new Map(active);
-        newActive.delete(stmt.name);
-        return { stmt: { ...stmt, initializer: init }, active: newActive };
-      }
-      return { stmt: { ...stmt, initializer: init }, active };
-    }
-    case "tuple-destructuring": {
-      const renamedInit = renameInExpression(stmt.initializer, active);
-      const newActive = new Map(active);
-      const newNames = stmt.names.map((name) => {
-        if (name === null) return null;
-        const rename = allRenames.get(name);
-        if (rename) {
-          // Apply configured rename and activate it for this scope
-          newActive.set(name, rename);
-          return rename;
-        }
-        // Suppress pre-activated rename when a tuple element re-declares the same name
-        // without a new rename configured in this scope.
-        if (active.has(name) && !allRenames.has(name)) {
-          newActive.delete(name);
-        }
-        return name;
-      });
-      return {
-        stmt: { ...stmt, names: newNames, initializer: renamedInit },
-        active: newActive,
-      };
-    }
-    case "if":
-      return {
-        stmt: {
-          ...stmt,
-          condition: renameInExpression(stmt.condition, active),
-          thenBody: scopeAwareRenameBlock(stmt.thenBody, active, allRenames),
-          elseBody: stmt.elseBody
-            ? scopeAwareRenameBlock(stmt.elseBody, active, allRenames)
-            : undefined,
-        },
-        active,
-      };
-    case "for": {
-      const forActive = new Map(active);
-      let init = stmt.initializer;
-      if (init) {
-        if (init.kind === "variable-declaration") {
-          const rename = allRenames.get(init.name);
-          if (rename) {
-            // Rename the initializer with the parent scope (before loop var is active)
-            const renamedInit = init.initializer
-              ? renameInExpression(init.initializer, forActive)
-              : undefined;
-            forActive.set(init.name, rename);
-            init = { ...init, name: rename, initializer: renamedInit };
-          } else {
-            const initName = init.name;
-            init = {
-              ...init,
-              initializer: init.initializer
-                ? renameInExpression(init.initializer, forActive)
-                : undefined,
-            } as typeof init;
-            // Suppress pre-activated rename when for-loop init re-declares the same name
-            if (forActive.has(initName) && !allRenames.has(initName)) {
-              forActive.delete(initName);
-            }
-          }
-        } else {
-          init = renameInStatement(init, forActive) as typeof init;
-        }
-      }
-      return {
-        stmt: {
-          ...stmt,
-          initializer: init,
-          condition: stmt.condition
-            ? renameInExpression(stmt.condition, forActive)
-            : undefined,
-          incrementor: stmt.incrementor
-            ? renameInExpression(stmt.incrementor, forActive)
-            : undefined,
-          body: scopeAwareRenameBlock(stmt.body, forActive, allRenames),
-        },
-        active,
-      };
-    }
-    case "while":
-      return {
-        stmt: {
-          ...stmt,
-          condition: renameInExpression(stmt.condition, active),
-          body: scopeAwareRenameBlock(stmt.body, active, allRenames),
-        },
-        active,
-      };
-    case "do-while":
-      return {
-        stmt: {
-          ...stmt,
-          body: scopeAwareRenameBlock(stmt.body, active, allRenames),
-          condition: renameInExpression(stmt.condition, active),
-        },
-        active,
-      };
-    case "switch":
-      return {
-        stmt: {
-          ...stmt,
-          discriminant: renameInExpression(stmt.discriminant, active),
-          cases: stmt.cases.map((c) => ({
-            ...c,
-            value: c.value ? renameInExpression(c.value, active) : undefined,
-            body: scopeAwareRenameBlock(c.body, active, allRenames),
-          })),
-        },
-        active,
-      };
-    case "try-catch": {
-      const tryActive = new Map(active);
-      let newReturnVarName = stmt.returnVarName;
-      if (stmt.returnVarName && allRenames.has(stmt.returnVarName)) {
-        const rename = allRenames.get(stmt.returnVarName)!;
-        tryActive.set(stmt.returnVarName, rename);
-        newReturnVarName = rename;
-      } else if (
-        stmt.returnVarName &&
-        tryActive.has(stmt.returnVarName) &&
-        !allRenames.has(stmt.returnVarName)
-      ) {
-        // Suppress pre-activated rename when try-catch return var re-declares the same name
-        tryActive.delete(stmt.returnVarName);
-      }
-      return {
-        stmt: {
-          ...stmt,
-          call: renameInExpression(stmt.call, active),
-          returnVarName: newReturnVarName,
-          successBody: scopeAwareRenameBlock(
-            stmt.successBody,
-            tryActive,
-            allRenames
-          ),
-          catchBody: scopeAwareRenameBlock(stmt.catchBody, active, allRenames),
-        },
-        active,
-      };
-    }
-    default:
-      return { stmt: renameInStatement(stmt, active), active };
-  }
-}
-
-export function resolveShadowedLocals(
-  body: Statement[],
-  stateVarNames: Set<string>,
-  paramNames?: Set<string>
-): Statement[] {
-  const localNames = collectLocalVarNames(body);
-  const shadowed = new Set<string>();
-  for (const name of localNames) {
-    if (stateVarNames.has(name)) {
-      shadowed.add(name);
-    }
-  }
-  if (shadowed.size === 0) return body;
-
-  const taken = new Set([
-    ...stateVarNames,
-    ...localNames,
-    ...(paramNames ?? []),
-  ]);
-  const renames = new Map<string, string>();
-  for (const name of shadowed) {
-    const newName = pickNewName(name, taken);
-    renames.set(name, newName);
-    taken.add(newName);
-  }
-  return scopeAwareRenameBlock(body, new Map(), renames);
-}
-
-function suffixToSolType(suffix: string): string {
-  if (suffix.startsWith("arr_")) return `${suffixToSolType(suffix.slice(4))}[]`;
-  return suffix;
-}
+export { generateType, generateExpression } from "./codegen/expressions.ts";
+export { generateStatement } from "./codegen/statements.ts";
+export { resolveShadowedLocals } from "./codegen/rename.ts";
+export { buildSourceMap } from "./codegen/sourcemap.ts";
 
 // ============================================================
 // Main entry
 // ============================================================
-
-/**
- * Compute the transitive set of ancestor contract names for the given contract
- * using a BFS traversal over `inherits`.
- */
-function computeAncestors(
-  contract: SkittlesContract,
-  contractByName: Map<string, SkittlesContract>
-): Set<string> {
-  const ancestors = new Set<string>();
-  const queue = [...contract.inherits.filter((n) => contractByName.has(n))];
-  let queueIndex = 0;
-  while (queueIndex < queue.length) {
-    const name = queue[queueIndex++]!;
-    if (ancestors.has(name)) continue;
-    ancestors.add(name);
-    const parent = contractByName.get(name);
-    if (parent) {
-      for (const gp of parent.inherits) {
-        if (contractByName.has(gp)) queue.push(gp);
-      }
-    }
-  }
-  return ancestors;
-}
 
 /**
  * Generate a Solidity file containing multiple contracts.
@@ -950,12 +188,9 @@ export function generateSolidity(
   return generateSolidityFile([contract], imports, solidityConfig);
 }
 
-const getFunctionKey = (f: SkittlesFunction): string => {
-  const paramTypes = f.parameters
-    .map((p) => (p.type ? generateType(p.type) : "unknown"))
-    .join(",");
-  return `${f.name}(${paramTypes})`;
-};
+// ============================================================
+// Contract body generation
+// ============================================================
 
 function emitNatSpecTag(
   lines: string[],
@@ -993,12 +228,6 @@ function generateNatSpecLines(
   if (natspec.returns) emitNatSpecTag(lines, indent, "return", natspec.returns);
   return lines;
 }
-
-const hasAncestorOrigin = (
-  origins: Set<string> | undefined,
-  ancestors: Set<string>
-): boolean =>
-  origins !== undefined && Array.from(origins).some((o) => ancestors.has(o));
 
 function generateContractBody(
   contract: SkittlesContract,
@@ -1288,171 +517,6 @@ function generateContractFunctions(
   }
 }
 
-function emitHelperFunctions(
-  parts: string[],
-  addOrigin: (map: Map<string, Set<string>>, key: string) => void,
-  functionOrigins: Map<string, Set<string>>,
-  ancestors: Set<string>
-): void {
-  const needsHelper = (name: string, flag: boolean): boolean =>
-    flag && !hasAncestorOrigin(functionOrigins.get(name), ancestors);
-
-  const emitHelper = (name: string, lines: string[]): void => {
-    addOrigin(functionOrigins, name);
-    parts.push("");
-    for (const line of lines) parts.push(line);
-  };
-
-  // Emit registered helpers using the table-driven registry
-  for (const entry of HELPER_REGISTRY) {
-    if (needsHelper(entry.name, cctx.helpers.has(entry.name))) {
-      emitHelper(entry.name, entry.lines());
-    }
-  }
-
-  // Emit array helper functions based on what methods are used,
-  // skipping any already emitted by an ancestor contract in this file
-  for (const helperKey of cctx.currentNeededArrayHelpers) {
-    const [method, ...typeParts] = helperKey.split("_");
-    const suffix = typeParts.join("_");
-    const solType = suffixToSolType(suffix);
-    const isRefType =
-      !SOLIDITY_VALUE_TYPES.has(solType) &&
-      !cctx.allKnownEnumNames.has(solType) &&
-      !cctx.allKnownInterfaceNames.has(solType);
-    const useHashEq = solType === "string" || solType === "bytes";
-    const memAnnotation = isRefType ? "memory " : "";
-    const eqCheck = useHashEq
-      ? `keccak256(abi.encodePacked(arr[i])) == keccak256(abi.encodePacked(value))`
-      : `arr[i] == value`;
-
-    if (method === "includes" && needsHelper(`_arrIncludes_${suffix}`, true)) {
-      emitHelper(`_arrIncludes_${suffix}`, [
-        `    function _arrIncludes_${suffix}(${solType}[] storage arr, ${solType} ${memAnnotation}value) internal view returns (bool) {`,
-        `        for (uint256 i = 0; i < arr.length; i++) {`,
-        `            if (${eqCheck}) return true;`,
-        `        }`,
-        `        return false;`,
-        `    }`,
-      ]);
-    }
-
-    if (method === "indexOf" && needsHelper(`_arrIndexOf_${suffix}`, true)) {
-      emitHelper(`_arrIndexOf_${suffix}`, [
-        `    function _arrIndexOf_${suffix}(${solType}[] storage arr, ${solType} ${memAnnotation}value) internal view returns (uint256) {`,
-        `        for (uint256 i = 0; i < arr.length; i++) {`,
-        `            if (${eqCheck}) return i;`,
-        `        }`,
-        `        return type(uint256).max;`,
-        `    }`,
-      ]);
-    }
-
-    if (
-      method === "lastIndexOf" &&
-      needsHelper(`_arrLastIndexOf_${suffix}`, true)
-    ) {
-      emitHelper(`_arrLastIndexOf_${suffix}`, [
-        `    function _arrLastIndexOf_${suffix}(${solType}[] storage arr, ${solType} ${memAnnotation}value) internal view returns (uint256) {`,
-        `        for (uint256 i = arr.length; i > 0; i--) {`,
-        `            if (${eqCheck.replace(/arr\[i\]/g, "arr[i - 1]")}) return i - 1;`,
-        `        }`,
-        `        return type(uint256).max;`,
-        `    }`,
-      ]);
-    }
-
-    if (method === "remove" && needsHelper(`_arrRemove_${suffix}`, true)) {
-      emitHelper(`_arrRemove_${suffix}`, [
-        `    function _arrRemove_${suffix}(${solType}[] storage arr, ${solType} ${memAnnotation}value) internal returns (bool) {`,
-        `        for (uint256 i = 0; i < arr.length; i++) {`,
-        `            if (${eqCheck}) {`,
-        `                arr[i] = arr[arr.length - 1];`,
-        `                arr.pop();`,
-        `                return true;`,
-        `            }`,
-        `        }`,
-        `        return false;`,
-        `    }`,
-      ]);
-    }
-
-    if (method === "reverse" && needsHelper(`_arrReverse_${suffix}`, true)) {
-      emitHelper(`_arrReverse_${suffix}`, [
-        `    function _arrReverse_${suffix}(${solType}[] storage arr) internal {`,
-        `        uint256 len = arr.length;`,
-        `        for (uint256 i = 0; i < len / 2; i++) {`,
-        `            ${solType} ${memAnnotation}temp = arr[i];`,
-        `            arr[i] = arr[len - 1 - i];`,
-        `            arr[len - 1 - i] = temp;`,
-        `        }`,
-        `    }`,
-      ]);
-    }
-
-    if (method === "splice" && needsHelper(`_arrSplice_${suffix}`, true)) {
-      emitHelper(`_arrSplice_${suffix}`, [
-        `    function _arrSplice_${suffix}(${solType}[] storage arr, uint256 start, uint256 deleteCount) internal {`,
-        `        require(start < arr.length, "${ERR_START_OUT_OF_BOUNDS}");`,
-        `        uint256 end = start + deleteCount;`,
-        `        if (end > arr.length) end = arr.length;`,
-        `        uint256 removed = end - start;`,
-        `        for (uint256 i = start; i < arr.length - removed; i++) {`,
-        `            arr[i] = arr[i + removed];`,
-        `        }`,
-        `        for (uint256 i = 0; i < removed; i++) {`,
-        `            arr.pop();`,
-        `        }`,
-        `    }`,
-      ]);
-    }
-
-    if (method === "slice" && needsHelper(`_arrSlice_${suffix}`, true)) {
-      emitHelper(`_arrSlice_${suffix}`, [
-        `    function _arrSlice_${suffix}(${solType}[] storage arr, uint256 start, uint256 end) internal view returns (${solType}[] memory) {`,
-        `        if (end > arr.length) end = arr.length;`,
-        `        require(start <= end, "invalid slice range");`,
-        `        ${solType}[] memory result = new ${solType}[](end - start);`,
-        `        for (uint256 i = start; i < end; i++) {`,
-        `            result[i - start] = arr[i];`,
-        `        }`,
-        `        return result;`,
-        `    }`,
-      ]);
-    }
-
-    if (method === "concat" && needsHelper(`_arrConcat_${suffix}`, true)) {
-      emitHelper(`_arrConcat_${suffix}`, [
-        `    function _arrConcat_${suffix}(${solType}[] storage arr, ${solType}[] memory other) internal view returns (${solType}[] memory) {`,
-        `        ${solType}[] memory result = new ${solType}[](arr.length + other.length);`,
-        `        for (uint256 i = 0; i < arr.length; i++) {`,
-        `            result[i] = arr[i];`,
-        `        }`,
-        `        for (uint256 i = 0; i < other.length; i++) {`,
-        `            result[arr.length + i] = other[i];`,
-        `        }`,
-        `        return result;`,
-        `    }`,
-      ]);
-    }
-
-    if (method === "spread" && needsHelper(`_arrSpread_${suffix}`, true)) {
-      emitHelper(`_arrSpread_${suffix}`, [
-        `    function _arrSpread_${suffix}(${solType}[] memory a, ${solType}[] memory b) internal pure returns (${solType}[] memory) {`,
-        `        ${solType}[] memory result = new ${solType}[](a.length + b.length);`,
-        `        for (uint256 i = 0; i < a.length; i++) {`,
-        `            result[i] = a[i];`,
-        `        }`,
-        `        for (uint256 i = 0; i < b.length; i++) {`,
-        `            result[a.length + i] = b[i];`,
-        `        }`,
-        `        return result;`,
-        `    }`,
-      ]);
-    }
-  }
-}
-
 // ============================================================
 // Contract elements
 // ============================================================
@@ -1648,7 +712,7 @@ function expressionMayModifyState(expr: Expression): boolean {
  *   - f(a, b) { return f(a, b, 10); }
  *   - f(a) { return f(a, 5, 10); }
  */
-function expandDefaultParamOverloads(f: SkittlesFunction): SkittlesFunction[] {
+export function expandDefaultParamOverloads(f: SkittlesFunction): SkittlesFunction[] {
   const defaultParams = f.parameters.filter((p) => p.defaultValue);
   if (defaultParams.length === 0) return [f];
 
@@ -1935,610 +999,6 @@ function generateConstructor(
 }
 
 // ============================================================
-// Type generation
-// ============================================================
-
-export function generateType(type: SkittlesType): string {
-  switch (type.kind) {
-    case SkittlesTypeKind.Uint256:
-      return "uint256";
-    case SkittlesTypeKind.Int256:
-      return "int256";
-    case SkittlesTypeKind.Address:
-      return "address";
-    case SkittlesTypeKind.Bool:
-      return "bool";
-    case SkittlesTypeKind.String:
-      return "string";
-    case SkittlesTypeKind.Bytes32:
-      return "bytes32";
-    case SkittlesTypeKind.Bytes:
-      return "bytes";
-    case SkittlesTypeKind.Mapping:
-      return `mapping(${generateType(type.keyType!)} => ${generateType(type.valueType!)})`;
-    case SkittlesTypeKind.Array:
-      return `${generateType(type.valueType!)}[]`;
-    case SkittlesTypeKind.Struct:
-      return type.structName ?? "UnknownStruct";
-    case SkittlesTypeKind.ContractInterface:
-      return type.structName ?? "UnknownInterface";
-    case SkittlesTypeKind.Enum:
-      return type.structName ?? "UnknownEnum";
-    case SkittlesTypeKind.Tuple:
-      return `(${(type.tupleTypes ?? []).map(generateType).join(", ")})`;
-    case SkittlesTypeKind.Void:
-      return "";
-    default: {
-      const _exhaustive: never = type.kind;
-      throw new Error(`Unhandled type kind: ${_exhaustive}`);
-    }
-  }
-}
-
-function generateParamType(type: SkittlesType): string {
-  const base = generateType(type);
-  if (needsMemoryLocation(type)) {
-    return `${base} memory`;
-  }
-  return base;
-}
-
-function generateCalldataParamType(type: SkittlesType): string {
-  const base = generateType(type);
-  if (needsMemoryLocation(type)) {
-    return `${base} calldata`;
-  }
-  return base;
-}
-
-function needsMemoryLocation(type: SkittlesType): boolean {
-  return [
-    SkittlesTypeKind.String,
-    SkittlesTypeKind.Bytes,
-    SkittlesTypeKind.Array,
-    SkittlesTypeKind.Struct,
-  ].includes(type.kind);
-}
-
-// ============================================================
-// Expression generation
-// ============================================================
-
-export function generateExpression(expr: Expression): string {
-  switch (expr.kind) {
-    case "number-literal":
-      return expr.value;
-    case "string-literal": {
-      if (ADDRESS_LITERAL_RE.test(expr.value)) {
-        return `address(${expr.value})`;
-      }
-      const escaped = expr.value
-        .replace(/\\/g, "\\\\")
-        .replace(/"/g, '\\"')
-        .replace(/[\x00-\x1f\x7f]/g, (ch) => {
-          const hex = ch.charCodeAt(0).toString(16).padStart(2, "0");
-          return `\\x${hex}`;
-        });
-      return `"${escaped}"`;
-    }
-    case "boolean-literal":
-      return expr.value ? "true" : "false";
-    case "identifier":
-      if (expr.name === "self") return "address(this)";
-      return expr.name;
-    case "property-access":
-      if (expr.object.kind === "identifier" && expr.object.name === "this") {
-        return expr.property;
-      }
-      // Number.MAX_VALUE → type(uint256).max
-      if (
-        expr.object.kind === "identifier" &&
-        expr.object.name === "Number" &&
-        expr.property === "MAX_VALUE"
-      ) {
-        return "type(uint256).max";
-      }
-      // Number.MAX_SAFE_INTEGER → 9007199254740991 (2^53 - 1)
-      if (
-        expr.object.kind === "identifier" &&
-        expr.object.name === "Number" &&
-        expr.property === "MAX_SAFE_INTEGER"
-      ) {
-        return "9007199254740991";
-      }
-      return `${generateExpression(expr.object)}.${expr.property}`;
-    case "element-access":
-      return `${generateExpression(expr.object)}[${generateExpression(expr.index)}]`;
-    case "binary":
-      return `(${generateExpression(expr.left)} ${expr.operator} ${generateExpression(expr.right)})`;
-    case "unary":
-      if (expr.prefix) {
-        return `${expr.operator}${generateExpression(expr.operand)}`;
-      }
-      return `${generateExpression(expr.operand)}${expr.operator}`;
-    case "assignment":
-      return `${generateExpression(expr.target)} ${expr.operator} ${generateExpression(expr.value)}`;
-    case "call": {
-      const callResult = tryGenerateBuiltinCall(expr);
-      if (callResult) return callResult;
-      // addr.transfer(amount) → payable(addr).transfer(amount)
-      // Exclude this.transfer(...) (internal call) and this.stateVar.transfer(...)
-      // (external contract interface call) to avoid misclassifying non-ETH transfers.
-      if (
-        expr.callee.kind === "property-access" &&
-        expr.callee.property === "transfer" &&
-        expr.args.length === 1 &&
-        !isThisOrContractCall(expr.callee.object)
-      ) {
-        const addr = generateExpression(expr.callee.object);
-        return `payable(${addr}).transfer(${generateExpression(expr.args[0])})`;
-      }
-      return `${generateExpression(expr.callee)}(${expr.args.map(generateExpression).join(", ")})`;
-    }
-    case "conditional":
-      return `(${generateExpression(expr.condition)} ? ${generateExpression(expr.whenTrue)} : ${generateExpression(expr.whenFalse)})`;
-    case "new":
-      return `new ${expr.callee}(${expr.args.map(generateExpression).join(", ")})`;
-    case "object-literal": {
-      const values = expr.properties
-        .map((p) => generateExpression(p.value))
-        .join(", ");
-      return values;
-    }
-    case "tuple-literal":
-      return `(${expr.elements.map(generateExpression).join(", ")})`;
-    default: {
-      const _exhaustive: never = expr;
-      throw new Error(`Unhandled expression kind: ${(_exhaustive as Expression).kind}`);
-    }
-  }
-}
-
-// ============================================================
-// Statement generation
-// ============================================================
-
-/**
- * Check whether an expression refers to an enum-typed state variable.
- * Used to decide whether a uint256(...) cast is needed in return statements.
- */
-function isEnumExpression(expr: Expression): boolean {
-  // Direct identifier: `status` (state variable)
-  if (expr.kind === "identifier") {
-    return cctx.currentEnumStateVarNames.has(expr.name);
-  }
-  // Property access: `this.status`
-  if (
-    expr.kind === "property-access" &&
-    expr.object.kind === "identifier" &&
-    expr.object.name === "this"
-  ) {
-    return cctx.currentEnumStateVarNames.has(expr.property);
-  }
-  return false;
-}
-
-export function generateStatement(stmt: Statement, indent: string): string {
-  const inner = indent + "    ";
-
-  switch (stmt.kind) {
-    case "return":
-      if (stmt.value) {
-        let retExpr = generateExpression(stmt.value);
-        // When the function returns uint256 but the expression is an enum
-        // state variable, emit an explicit uint256(...) cast so Solidity
-        // doesn't reject the implicit enum→uint256 conversion.
-        if (
-          cctx.currentFunctionReturnType?.kind === SkittlesTypeKind.Uint256 &&
-          isEnumExpression(stmt.value)
-        ) {
-          retExpr = `uint256(${retExpr})`;
-        }
-        return `${indent}return ${retExpr};`;
-      }
-      return `${indent}return;`;
-
-    case "variable-declaration": {
-      const type = stmt.type ? generateParamType(stmt.type) : "uint256";
-      if (stmt.initializer) {
-        let initExpr = generateExpression(stmt.initializer);
-        // Wrap object literal in struct constructor when the type is a struct
-        if (
-          stmt.initializer.kind === "object-literal" &&
-          stmt.type?.kind === SkittlesTypeKind.Struct
-        ) {
-          initExpr = `${stmt.type.structName}(${initExpr})`;
-        }
-        return `${indent}${type} ${stmt.name} = ${initExpr};`;
-      }
-      return `${indent}${type} ${stmt.name};`;
-    }
-
-    case "tuple-destructuring": {
-      const parts = stmt.names.map((name, i) => {
-        if (name === null) return "";
-        const elemType = i < stmt.types.length ? stmt.types[i] : null;
-        const type =
-          elemType !== null ? generateParamType(elemType) : "uint256";
-        return `${type} ${name}`;
-      });
-      const initExpr = generateExpression(stmt.initializer);
-      return `${indent}(${parts.join(", ")}) = ${initExpr};`;
-    }
-
-    case "expression": {
-      if (stmt.expression.kind === "conditional" && indent) {
-        const conditionalExpr = stmt.expression;
-        const lines: string[] = [];
-        lines.push(
-          `${indent}if (${generateExpression(conditionalExpr.condition)}) {`
-        );
-        lines.push(
-          generateStatement(
-            { kind: "expression", expression: conditionalExpr.whenTrue },
-            inner
-          )
-        );
-        lines.push(`${indent}} else {`);
-        lines.push(
-          generateStatement(
-            { kind: "expression", expression: conditionalExpr.whenFalse },
-            inner
-          )
-        );
-        lines.push(`${indent}}`);
-        return lines.join("\n");
-      }
-      return `${indent}${generateExpression(stmt.expression)};`;
-    }
-
-    case "if": {
-      if (isRequirePattern(stmt)) {
-        const negated = negateExpression(stmt.condition);
-        const revert = stmt.thenBody[0] as RevertStatement;
-        if (revert.message) {
-          return `${indent}require(${generateExpression(negated)}, ${generateExpression(revert.message)});`;
-        }
-        return `${indent}require(${generateExpression(negated)});`;
-      }
-
-      const lines: string[] = [];
-      lines.push(`${indent}if (${generateExpression(stmt.condition)}) {`);
-      for (const s of stmt.thenBody) {
-        lines.push(generateStatement(s, inner));
-      }
-      if (stmt.elseBody) {
-        lines.push(`${indent}} else {`);
-        for (const s of stmt.elseBody) {
-          lines.push(generateStatement(s, inner));
-        }
-      }
-      lines.push(`${indent}}`);
-      return lines.join("\n");
-    }
-
-    case "for": {
-      let initStr = "";
-      if (stmt.initializer) {
-        initStr = generateStatement(stmt.initializer, "").trimStart();
-        if (initStr.endsWith(";")) {
-          initStr = initStr.slice(0, -1);
-        }
-      }
-      const condStr = stmt.condition ? generateExpression(stmt.condition) : "";
-      const incrStr = stmt.incrementor
-        ? generateExpression(stmt.incrementor)
-        : "";
-
-      const lines: string[] = [];
-      lines.push(`${indent}for (${initStr}; ${condStr}; ${incrStr}) {`);
-      for (const s of stmt.body) {
-        lines.push(generateStatement(s, inner));
-      }
-      lines.push(`${indent}}`);
-      return lines.join("\n");
-    }
-
-    case "while": {
-      const lines: string[] = [];
-      lines.push(`${indent}while (${generateExpression(stmt.condition)}) {`);
-      for (const s of stmt.body) {
-        lines.push(generateStatement(s, inner));
-      }
-      lines.push(`${indent}}`);
-      return lines.join("\n");
-    }
-
-    case "do-while": {
-      const lines: string[] = [];
-      lines.push(`${indent}do {`);
-      for (const s of stmt.body) {
-        lines.push(generateStatement(s, inner));
-      }
-      lines.push(`${indent}} while (${generateExpression(stmt.condition)});`);
-      return lines.join("\n");
-    }
-
-    case "break":
-      return `${indent}break;`;
-
-    case "continue":
-      return `${indent}continue;`;
-
-    case "revert":
-      if (stmt.customError) {
-        const args = (stmt.customErrorArgs ?? [])
-          .map(generateExpression)
-          .join(", ");
-        return `${indent}revert ${stmt.customError}(${args});`;
-      }
-      if (stmt.message) {
-        return `${indent}revert(${generateExpression(stmt.message)});`;
-      }
-      return `${indent}revert();`;
-
-    case "emit":
-      return `${indent}emit ${stmt.eventName}(${stmt.args.map(generateExpression).join(", ")});`;
-
-    case "delete":
-      return `${indent}delete ${generateExpression(stmt.target)};`;
-
-    case "switch": {
-      const lines: string[] = [];
-      const discExpr = generateExpression(stmt.discriminant);
-      let first = true;
-      let defaultCase: Statement[] | undefined;
-
-      for (const c of stmt.cases) {
-        if (!c.value) {
-          defaultCase = c.body;
-          continue;
-        }
-        const keyword = first ? "if" : "} else if";
-        lines.push(
-          `${indent}${keyword} (${discExpr} == ${generateExpression(c.value)}) {`
-        );
-        for (const s of c.body) {
-          lines.push(generateStatement(s, inner));
-        }
-        first = false;
-      }
-
-      if (defaultCase) {
-        if (!first) {
-          lines.push(`${indent}} else {`);
-        } else {
-          lines.push(`${indent}{`);
-        }
-        for (const s of defaultCase) {
-          lines.push(generateStatement(s, inner));
-        }
-      }
-      lines.push(`${indent}}`);
-      return lines.join("\n");
-    }
-
-    case "try-catch": {
-      const lines: string[] = [];
-      const callExpr = generateExpression(stmt.call);
-      let returns = "";
-      if (
-        stmt.returnVarName &&
-        stmt.returnType &&
-        stmt.returnType.kind !== SkittlesTypeKind.Void
-      ) {
-        returns = ` returns (${generateType(stmt.returnType)} ${stmt.returnVarName})`;
-      }
-      lines.push(`${indent}try ${callExpr}${returns} {`);
-      for (const s of stmt.successBody) {
-        lines.push(generateStatement(s, inner));
-      }
-      lines.push(`${indent}} catch {`);
-      for (const s of stmt.catchBody) {
-        lines.push(generateStatement(s, inner));
-      }
-      lines.push(`${indent}}`);
-      return lines.join("\n");
-    }
-
-    case "console-log":
-      return `${indent}console.log(${stmt.args.map(generateExpression).join(", ")});`;
-
-    default: {
-      const _exhaustive: never = stmt;
-      throw new Error(`Unhandled statement kind: ${(_exhaustive as Statement).kind}`);
-    }
-  }
-}
-
-// ============================================================
-// Require pattern optimization
-// ============================================================
-
-function isRequirePattern(stmt: IfStatement): boolean {
-  if (
-    stmt.thenBody.length === 1 &&
-    stmt.thenBody[0].kind === "revert" &&
-    !stmt.elseBody
-  ) {
-    const revert = stmt.thenBody[0];
-    if (revert.customError) return false;
-    return true;
-  }
-  return false;
-}
-
-function negateExpression(expr: Expression): Expression {
-  if (expr.kind === "binary") {
-    const negated = negateOperator(expr.operator);
-    if (negated) {
-      return { ...expr, operator: negated };
-    }
-  }
-  return { kind: "unary", operator: "!", operand: expr, prefix: true };
-}
-
-function negateOperator(op: string): string | null {
-  const map: Record<string, string> = {
-    "<": ">=",
-    ">": "<=",
-    "<=": ">",
-    ">=": "<",
-    "==": "!=",
-    "!=": "==",
-  };
-  return map[op] ?? null;
-}
-
-/**
- * Check if a receiver expression is `this` (an internal contract method call).
- * Used to avoid wrapping `this.transfer(...)` calls in `payable(...)`.
- * Note: `this.stateVar.transfer(amount)` is NOT excluded here because
- * codegen strips `this.` and the arg count (1 for ETH transfer vs 2+ for
- * contract interface calls) serves as the discriminator.
- */
-function isThisOrContractCall(receiver: Expression): boolean {
-  return receiver.kind === "identifier" && receiver.name === "this";
-}
-
-// ============================================================
-// Built-in function recognition
-// ============================================================
-
-function getCallName(expr: Expression): string | null {
-  if (expr.kind === "identifier") return expr.name;
-  if (expr.kind === "property-access" && expr.object.kind === "identifier") {
-    return `${expr.object.name}.${expr.property}`;
-  }
-  return null;
-}
-
-function tryGenerateBuiltinCall(expr: {
-  callee: Expression;
-  args: Expression[];
-  typeArgs?: SkittlesType[];
-}): string | null {
-  const name = getCallName(expr.callee);
-  if (!name) return null;
-  const args = expr.args.map(generateExpression).join(", ");
-
-  switch (name) {
-    case "hash":
-    case "keccak256":
-      return `keccak256(abi.encodePacked(${args}))`;
-    case "sha256":
-      return `sha256(abi.encodePacked(${args}))`;
-    case "abi.encode":
-      return `abi.encode(${args})`;
-    case "abi.encodePacked":
-      return `abi.encodePacked(${args})`;
-    case "abi.decode": {
-      if (expr.typeArgs && expr.typeArgs.length > 0) {
-        const types = expr.typeArgs.map(generateType).join(", ");
-        return `abi.decode(${args}, (${types}))`;
-      }
-      return `abi.decode(${args})`;
-    }
-    case "ecrecover": {
-      const [hashArg, vArg, rArg, sArg] = expr.args.map(generateExpression);
-      return `ecrecover(${hashArg}, uint8(${vArg}), ${rArg}, ${sArg})`;
-    }
-    case "addmod":
-      return `addmod(${args})`;
-    case "mulmod":
-      return `mulmod(${args})`;
-    case "assert":
-      return `assert(${args})`;
-    case "gasleft":
-      return `gasleft()`;
-    case "Contract":
-      if (
-        expr.typeArgs &&
-        expr.typeArgs.length > 0 &&
-        expr.typeArgs[0].kind === SkittlesTypeKind.ContractInterface &&
-        expr.typeArgs[0].structName
-      ) {
-        return `${expr.typeArgs[0].structName}(${args})`;
-      }
-      throw new Error(
-        "Contract<T>() requires a contract interface type argument, e.g. Contract<IToken>(address)"
-      );
-    case "string.concat":
-      return `string.concat(${args})`;
-    case "bytes.concat":
-      return `bytes.concat(${args})`;
-    case "__sk_toString": {
-      cctx.helpers.add("__sk_toString");
-      return `__sk_toString(${args})`;
-    }
-    case "Math.min": {
-      cctx.helpers.add("_min");
-      const a = generateExpression(expr.args[0]);
-      const b = generateExpression(expr.args[1]);
-      return `_min(${a}, ${b})`;
-    }
-    case "Math.max": {
-      cctx.helpers.add("_max");
-      const a = generateExpression(expr.args[0]);
-      const b = generateExpression(expr.args[1]);
-      return `_max(${a}, ${b})`;
-    }
-    case "Math.pow": {
-      const base = generateExpression(expr.args[0]);
-      const exp = generateExpression(expr.args[1]);
-      return `(${base} ** ${exp})`;
-    }
-    case "Math.sqrt": {
-      cctx.helpers.add("_sqrt");
-      const x = generateExpression(expr.args[0]);
-      return `_sqrt(${x})`;
-    }
-    case "_charAt": {
-      cctx.helpers.add("_charAt");
-      return `_charAt(${args})`;
-    }
-    case "_substring": {
-      cctx.helpers.add("_substring");
-      return `_substring(${args})`;
-    }
-    case "_toLowerCase": {
-      cctx.helpers.add("_toLowerCase");
-      return `_toLowerCase(${args})`;
-    }
-    case "_toUpperCase": {
-      cctx.helpers.add("_toUpperCase");
-      return `_toUpperCase(${args})`;
-    }
-    case "_startsWith": {
-      cctx.helpers.add("_startsWith");
-      return `_startsWith(${args})`;
-    }
-    case "_endsWith": {
-      cctx.helpers.add("_endsWith");
-      return `_endsWith(${args})`;
-    }
-    case "_trim": {
-      cctx.helpers.add("_trim");
-      return `_trim(${args})`;
-    }
-    case "_split": {
-      cctx.helpers.add("_split");
-      return `_split(${args})`;
-    }
-    case "_replace": {
-      cctx.helpers.add("_replace");
-      return `_replace(${args})`;
-    }
-    case "_replaceAll": {
-      cctx.helpers.add("_replaceAll");
-      return `_replaceAll(${args})`;
-    }
-    default:
-      return null;
-  }
-}
-
-// ============================================================
 // Console.log detection
 // ============================================================
 
@@ -2567,192 +1027,4 @@ function contractUsesConsoleLog(contract: SkittlesContract): boolean {
 function mapVisibility(vis: Visibility): string {
   if (vis === "public") return "public";
   return "internal";
-}
-
-// ============================================================
-// Source map generation
-// ============================================================
-
-/**
- * Build a source map that maps generated Solidity line numbers
- * back to TypeScript source line numbers.
- *
- * The mapping is built by walking the IR (which has source line info
- * from the parser) and counting lines in the generated Solidity output
- * to correlate each Solidity line with its TypeScript origin.
- */
-export function buildSourceMap(
-  solidity: string,
-  contracts: SkittlesContract[],
-  sourceFile: string
-): SourceMapping {
-  const solLines = solidity.split("\n");
-  const mappings: Record<number, number> = {};
-
-  let lineIdx = 0; // 0-based index into solLines
-
-  // Helper: find the next line matching a test, starting from lineIdx
-  function findLine(test: (line: string) => boolean): number {
-    for (let i = lineIdx; i < solLines.length; i++) {
-      if (test(solLines[i])) {
-        lineIdx = i;
-        return i;
-      }
-    }
-    return -1;
-  }
-
-  function addMapping(solLineIdx: number, tsLine: number | undefined): void {
-    if (tsLine !== undefined && solLineIdx >= 0) {
-      mappings[solLineIdx + 1] = tsLine; // convert to 1-based
-    }
-  }
-
-  /**
-   * Map function/constructor body statements to Solidity lines.
-   * Walks statements in order, using generateStatement to count lines
-   * for each statement so we know exactly where each one appears.
-   */
-  function mapBodyStatements(
-    body: Statement[],
-    startLineIdx: number,
-    indent: string
-  ): void {
-    let currentIdx = startLineIdx;
-    for (const stmt of body) {
-      addMapping(currentIdx, stmt.sourceLine);
-      const stmtText = generateStatement(stmt, indent);
-      const stmtLineCount = stmtText.split("\n").length;
-
-      // Recurse into compound statement bodies
-      if (stmt.kind === "if" && !isRequirePattern(stmt)) {
-        // Line 0: if (cond) {
-        mapBodyStatements(stmt.thenBody, currentIdx + 1, indent + "    ");
-        if (stmt.elseBody) {
-          const thenLineCount = stmt.thenBody.reduce(
-            (sum, s) =>
-              sum + generateStatement(s, indent + "    ").split("\n").length,
-            0
-          );
-          // } else { is at currentIdx + 1 + thenLineCount
-          mapBodyStatements(
-            stmt.elseBody,
-            currentIdx + 1 + thenLineCount + 1,
-            indent + "    "
-          );
-        }
-      } else if (
-        stmt.kind === "expression" &&
-        stmt.expression.kind === "conditional" &&
-        indent
-      ) {
-        // Lowered void ternary: map all generated lines to the original source line
-        for (let i = 1; i < stmtLineCount; i++) {
-          addMapping(currentIdx + i, stmt.sourceLine);
-        }
-      } else if (stmt.kind === "for" || stmt.kind === "while") {
-        mapBodyStatements(stmt.body, currentIdx + 1, indent + "    ");
-      } else if (stmt.kind === "do-while") {
-        mapBodyStatements(stmt.body, currentIdx + 1, indent + "    ");
-      }
-
-      currentIdx += stmtLineCount;
-    }
-  }
-
-  // Replicate the same ancestor-origin suppression used during codegen so we
-  // only try to map functions that were actually emitted in the Solidity output.
-  const contractByName = new Map(contracts.map((c) => [c.name, c] as const));
-  const smAncestorsMap = new Map<string, Set<string>>();
-  for (const c of contracts) {
-    smAncestorsMap.set(c.name, computeAncestors(c, contractByName));
-  }
-  const smFunctionOrigins = new Map<string, Set<string>>();
-
-  for (const contract of contracts) {
-    const smAncestors = smAncestorsMap.get(contract.name) ?? new Set<string>();
-
-    const functionsToMap = contract.functions.filter((f) => {
-      const key = getFunctionKey(f);
-      return !hasAncestorOrigin(smFunctionOrigins.get(key), smAncestors) || f.isOverride;
-    });
-    for (const f of functionsToMap) {
-      const key = getFunctionKey(f);
-      let origins = smFunctionOrigins.get(key);
-      if (!origins) {
-        origins = new Set<string>();
-        smFunctionOrigins.set(key, origins);
-      }
-      origins.add(contract.name);
-    }
-
-    // Find the contract declaration line
-    const contractIdx = findLine((l) => {
-      const trimmed = l.trimStart();
-      return (
-        trimmed.startsWith(`contract ${contract.name}`) ||
-        trimmed.startsWith(`abstract contract ${contract.name}`)
-      );
-    });
-    if (contractIdx === -1) continue;
-    addMapping(contractIdx, contract.sourceLine);
-    lineIdx = contractIdx + 1;
-
-    // Map events
-    for (const e of contract.events) {
-      const idx = findLine((l) => {
-        const trimmed = l.trim();
-        return trimmed.startsWith(`event ${e.name}(`);
-      });
-      if (idx !== -1) {
-        addMapping(idx, e.sourceLine);
-        lineIdx = idx + 1;
-      }
-    }
-
-    // Map variables
-    for (const v of contract.variables) {
-      const idx = findLine((l) => {
-        const trimmed = l.trim();
-        return trimmed.includes(` ${v.name}`) && trimmed.endsWith(";");
-      });
-      if (idx !== -1) {
-        addMapping(idx, v.sourceLine);
-        lineIdx = idx + 1;
-      }
-    }
-
-    // Map constructor
-    if (contract.ctor) {
-      const ctorIdx = findLine((l) => l.trim().startsWith("constructor("));
-      if (ctorIdx !== -1) {
-        addMapping(ctorIdx, contract.ctor.sourceLine);
-        lineIdx = ctorIdx + 1;
-        mapBodyStatements(contract.ctor.body, lineIdx, "        ");
-      }
-    }
-
-    // Map functions — expand default-param overloads so the source map
-    // scanner advances past overload wrapper bodies that appear in the
-    // generated Solidity (the same expansion is applied during codegen).
-    const expandedFns: SkittlesFunction[] = [];
-    for (const f of functionsToMap) {
-      expandedFns.push(...expandDefaultParamOverloads(f));
-    }
-    for (const f of expandedFns) {
-      const funcIdx = findLine((l) => {
-        const trimmed = l.trim();
-        if (f.name === "receive") return trimmed.startsWith("receive()");
-        if (f.name === "fallback") return trimmed.startsWith("fallback()");
-        return trimmed.startsWith(`function ${f.name}(`);
-      });
-      if (funcIdx !== -1) {
-        addMapping(funcIdx, f.sourceLine);
-        lineIdx = funcIdx + 1;
-        mapBodyStatements(f.body, lineIdx, "        ");
-      }
-    }
-  }
-
-  return { sourceFile, mappings };
 }
